Add logic to command buttons to permit prepend and append
Added buttonStart to the command

Added prepend logic to set cursor right before the button. This permits any output from the command to be prepended to the button.

Added append logic to set cursor right after the button (and after the button block id). This permits any output from the command to be appended to the button.Added buttonStart to the command

Added prepend logic to set cursor right before the button. This permits any output from the command to be prepended to the button.

Added append logic to set cursor right after the button (and after the button block id). This permits any output from the command to be appended to the button.