Add Prepend and Append logic to Command buttons (#169)
* Add buttonPosition to permit prepend and append

Add the buttonStart variable and pass it to the command.

* Add logic to command buttons to permit prepend and append

Added buttonStart to the command

Added prepend logic to set cursor right before the button. This permits any output from the command to be prepended to the button.

Added append logic to set cursor right after the button (and after the button block id). This permits any output from the command to be appended to the button.Added buttonStart to the command

Added prepend logic to set cursor right before the button. This permits any output from the command to be prepended to the button.

Added append logic to set cursor right after the button (and after the button block id). This permits any output from the command to be appended to the button.

* Update modal to accept options for command

Adding a Default, a Prepend, and an Append option when creating a command button. Also storing that value.

---------

Co-authored-by: Sam <sam@sbgood.ca>